Convert to int to avoid warning during compilation

+++ gam_g4/gam_g4/gam_lib/GamSPSVoxelsPosDistribution.cpp
@@ -35,7 +35,7 @@ G4ThreeVector GamSPSVoxelsPosDistribution::VGenerateOne() {
         auto p = G4UniformRand();
         auto lower = std::lower_bound(fCDFZ.begin(), fCDFZ.end(), p);
         i = std::distance(fCDFZ.begin(), lower);
-    } while(i >= fCDFX.size());
+    } while(i >= (int)fCDFX.size());
 
     // Get Cumulative Distribution Function for Y, knowing Z
     auto j = 0;
@@ -43,7 +43,7 @@ G4ThreeVector GamSPSVoxelsPosDistribution::VGenerateOne() {
         auto p = G4UniformRand();
         auto lower = std::lower_bound(fCDFY[i].begin(), fCDFY[i].end(), p);
         j = std::distance(fCDFY[i].begin(), lower);
-    } while(j >= fCDFX[i].size());
+    } while(j >= (int)fCDFX[i].size());
 
     // Get Cumulative Distribution Function for X, knowing X and Y
     auto k = 0;
@@ -51,7 +51,7 @@ G4ThreeVector GamSPSVoxelsPosDistribution::VGenerateOne() {
         auto p = G4UniformRand();
         auto lower = std::lower_bound(fCDFX[i][j].begin(), fCDFX[i][j].end(), p);
         k = std::distance(fCDFX[i][j].begin(), lower);
-    } while(k >= fCDFX[i][j].size());
+    } while(k >= (int)fCDFX[i][j].size());
 
     // convert to physical coordinate
     // (warning to the numpy order Z Y X)
